refactored modal to be dynamic and reusable so that i can use for both properties as well as tasks. added ability to delete task and rediract to edit task page for specific tasks
