+++ app/properties/page.jsx
@@ -6,7 +6,7 @@ import Nav from '../components/Nav';
 import styles from '../../styles/properties.module.css'
 import Loading from '../components/Loading';
 import {useDisclosure} from "@nextui-org/react";
-import PropertyModal from '../components/PropertyModal';
+import DetailsModal from '../components/DetailsModal';
 import Link from 'next/link';
 import { MdOutlineAddHome } from "react-icons/md";
 
@@ -209,7 +209,7 @@ const Properties = () => {
                             ))}
                         </ul>
                         {/* Passing down props to display property details in separate modal component */}
-                        <PropertyModal isOpen={isOpen} onClose={onOpenChange} property={selectedProperty} />
+                        <DetailsModal isOpen={isOpen} onClose={onOpenChange} data={selectedProperty} />
                     </div>
                 </div>
             </div>
